client/gemini/weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.11.0",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]
